refactor: move model/engine into bot folder
