+++ bot/model.py
